Move getLayout to dash_component_api
