+++ dash/dash-renderer/src/utils/stores.ts
@@ -1,0 +1,5 @@
+export function getStores() {
+    const stores = ((window as any).dash_stores =
+        (window as any).dash_stores || []);
+    return stores;
+}
